전체 페이지 API 연동 및 프론트 상태 처리와 경로 재구성 등
