+++ PR_DESCRIPTION.md
@@ -1,0 +1,88 @@
+# PR 제목
+```
+feat: 도메인 경로 재구성 및 UI 요소 숨김 처리
+```
+
+# PR 본문
+
+## 📋 변경 사항 요약
+
+이번 PR은 도메인 경로 구조를 재구성하고, 일부 UI 요소를 숨김 처리하는 작업을 포함합니다.
+
+## 🎯 주요 변경사항
+
+### 1. 도메인 경로 재구성
+- **온보딩 가이드**: `/onboarding/guide` → `/onboarding`
+- **온보딩 폼**: `/onboarding` → `/register`
+- **전체 통화 목록**: `/call-history` → `/detail`
+- **통화 상세**: `/call-history/[id]` → `/detail/[id]`
+
+### 2. 파일 구조 변경
+```
+app/
+├── onboarding/
+│   └── page.tsx          (온보딩 가이드 - 기존 onboarding/guide)
+├── register/
+│   ├── page.tsx          (온보딩 폼 - 기존 onboarding)
+│   ├── complete/
+│   │   └── page.tsx
+│   └── layout.tsx
+└── (dashboard)/
+    └── detail/
+        ├── page.tsx      (통화 목록 - 기존 call-history)
+        └── [id]/
+            └── page.tsx  (통화 상세)
+```
+
+### 3. UI 요소 숨김 처리
+- 사이드바에서 '소리 맞춤 설정' 메뉴 숨김 처리
+- 대시보드 헤더 우측 상단의 알림 아이콘 숨김 처리
+
+### 4. 경로 참조 업데이트
+다음 파일들의 라우팅 경로를 업데이트했습니다:
+- `app/page.tsx`
+- `app/dashboard/page.tsx`
+- `app/components/LNB.tsx`
+- `app/components/Sidebar.tsx`
+- `app/(dashboard)/detail/page.tsx`
+
+## 🔍 변경된 파일 목록
+
+### 수정된 파일
+- `app/components/LNB.tsx` - 사이드바 설정 메뉴 숨김, 경로 업데이트
+- `app/components/Sidebar.tsx` - 사이드바 설정 메뉴 숨김, 경로 업데이트
+- `app/dashboard/page.tsx` - 알림 숨김, 경로 업데이트
+- `app/page.tsx` - 경로 업데이트
+- `app/onboarding/page.tsx` - 경로 업데이트
+
+### 이동된 파일
+- `app/onboarding/page.tsx` (기존: `app/onboarding/guide/page.tsx`)
+- `app/register/*` (기존: `app/onboarding/*`)
+- `app/(dashboard)/detail/*` (기존: `app/(dashboard)/call-history/*`)
+
+### 삭제된 파일
+- `app/register/guide/page.tsx` (중복 파일)
+- `app/onboarding/guide/page.tsx` (이동됨)
+- `app/onboarding/complete/page.tsx` (이동됨)
+- `app/onboarding/layout.tsx` (이동됨)
+- `app/(dashboard)/call-history/*` (이동됨)
+
+## ✅ 테스트 체크리스트
+
+- [ ] 온보딩 가이드 페이지 접근 확인 (`/onboarding`)
+- [ ] 온보딩 폼 페이지 접근 확인 (`/register`)
+- [ ] 통화 목록 페이지 접근 확인 (`/detail`)
+- [ ] 통화 상세 페이지 접근 확인 (`/detail/[id]`)
+- [ ] 사이드바에서 설정 메뉴 미표시 확인
+- [ ] 대시보드 헤더에서 알림 미표시 확인
+- [ ] 기존 기능 정상 동작 확인
+
+## 📝 참고사항
+
+- 모든 경로 변경은 화면 표시에만 영향을 주며, 기존 기능은 그대로 유지됩니다.
+- UI 요소 숨김 처리는 조건부 렌더링을 통해 구현되어 향후 쉽게 복구 가능합니다.
+- 다른 디자인 및 UX 요소는 변경하지 않았습니다.
+
+## 🔗 관련 이슈
+- Issue #7
+
