+++ app/(dashboard)/detail/[id]/page.tsx
