+++ app/components/LNB.tsx
@@ -38,7 +38,7 @@ export function LNB() {
           <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M3 5a2 2 0 012-2h3.28a1 1 0 01.948.684l1.498 4.493a1 1 0 01-.502 1.21l-2.257 1.13a11.042 11.042 0 005.516 5.516l1.13-2.257a1 1 0 011.21-.502l4.493 1.498a1 1 0 01.684.949V19a2 2 0 01-2 2h-1C9.716 21 3 14.284 3 6V5z" />
         </svg>
       ),
-      path: '/call-history'
+      path: '/detail'
     },
     {
       id: 'chat',
@@ -95,7 +95,7 @@ export function LNB() {
         <SidebarGroup>
           <SidebarGroupContent>
             <SidebarMenu className="space-y-2">
-              {navItems.map((item) => {
+              {navItems.filter(item => item.id !== 'settings').map((item) => {
                 const isActive = pathname === item.path || pathname?.startsWith(item.path + '/');
                 return (
                   <SidebarMenuItem key={item.id}>
